Fork core/editor store for native instead of modifying it
