+++ packages/editor/src/store/actions.js
@@ -738,20 +738,6 @@ export function updatePostLock( lock ) {
 }
 
 /**
- * Returns an action object that enables or disables post title selection.
- *
- * @param {boolean} [isSelected=true] Whether post title is currently selected.
-
- * @return {Object} Action object.
- */
-export function togglePostTitleSelection( isSelected = true ) {
-	return {
-		type: 'TOGGLE_POST_TITLE_SELECTION',
-		isSelected,
-	};
-}
-
-/**
  * Returns an action object used to fetch a single reusable block or all
  * reusable blocks from the REST API into the store.
  *
