+++ packages/editor/src/store/actions.native.js
@@ -1,0 +1,16 @@
+
+export * from './actions.js';
+
+/**
+ * Returns an action object that enables or disables post title selection.
+ *
+ * @param {boolean} [isSelected=true] Whether post title is currently selected.
+
+ * @return {Object} Action object.
+ */
+export function togglePostTitleSelection( isSelected = true ) {
+	return {
+		type: 'TOGGLE_POST_TITLE_SELECTION',
+		isSelected,
+	};
+}
